+++ packages/apps/stacc-website/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [8.81.0](https://bitbucket.org/stacc-flow/bento/compare/v8.80.1...v8.81.0) (2019-03-28)
+
+**Note:** Version bump only for package @staccx/stacc-website
+
+
+
+
+
 ## [8.80.1](https://bitbucket.org/stacc-flow/bento/compare/v8.80.0...v8.80.1) (2019-03-27)